Slightly re-organize the pipeline code
This commit is a mostly mechanical shuffle of our pipeline code, done
while investigating #10144. I found the previous organization
challenging to comprehend and this reads more clear to me. I've tried
to follow the section commenting style common to the rest of the
transform code.
